+++ src/features/features.service.ts
@@ -1,6 +1,7 @@
 import {
     BadRequestException,
     Injectable,
+    InternalServerErrorException,
     NotFoundException,
 } from "@nestjs/common";
 import { PrismaService } from "../prisma/prisma.service";
@@ -8,6 +9,7 @@ import { CreateFeatureDto } from "./dto/create-feature.dto";
 import { UpdateFeatureDto } from "./dto/update-feature.dto";
 import { UpdateFeatureOrderAndCategoryDto } from "./dto/update-feature-order-and-category.dto";
 import { GlobalService } from "../global/global.service";
+import { CustomRequest } from "../global/types/CustomRequest";
 
 @Injectable()
 export class FeaturesService {
@@ -17,7 +19,7 @@ export class FeaturesService {
     ) {}
 
     async createFeature(
-        req,
+        req: CustomRequest,
         teamId: number,
         createFeatureDto: CreateFeatureDto,
     ) {
@@ -48,7 +50,8 @@ export class FeaturesService {
                 },
             });
 
-            const newOrder = lastFeature ? lastFeature.order + 1 : 1;
+            const newOrder =
+                lastFeature && lastFeature?.order ? lastFeature.order! + 1 : 1;
 
             const newFeature = await this.prisma.projectFeature.create({
                 data: {
@@ -79,6 +82,18 @@ export class FeaturesService {
 
     async findAllFeatures(teamId: number) {
         try {
+            //check for valid teamId
+            const team = await this.prisma.voyageTeam.findFirst({
+                where: {
+                    id: teamId,
+                },
+            });
+
+            if (!team) {
+                throw new NotFoundException(
+                    `TeamId (id: ${teamId}) does not exist.`,
+                );
+            }
             const allTeamFeatures = await this.prisma.projectFeature.findMany({
                 where: {
                     addedBy: {
@@ -113,12 +128,6 @@ export class FeaturesService {
                 },
                 orderBy: [{ category: { id: "asc" } }, { order: "asc" }],
             });
-
-            if (!allTeamFeatures) {
-                throw new NotFoundException(
-                    `TeamId (id: ${teamId}) does not exist.`,
-                );
-            }
 
             return allTeamFeatures;
         } catch (e) {
@@ -193,7 +202,7 @@ export class FeaturesService {
     }
 
     async updateFeatureOrderAndCategory(
-        req,
+        req: CustomRequest,
         featureId: number,
         updateOrderAndCategoryDto: UpdateFeatureOrderAndCategoryDto,
     ) {
@@ -202,11 +211,17 @@ export class FeaturesService {
 
         const currFeature = await this.findFeature(featureId);
 
-        const { voyageTeamId: teamId } =
-            await this.prisma.voyageTeamMember.findFirst({
-                where: { id: currFeature.teamMemberId },
-                select: { voyageTeamId: true },
-            });
+        const voyageTeamMember = await this.prisma.voyageTeamMember.findFirst({
+            where: { id: currFeature.teamMemberId! },
+            select: { voyageTeamId: true },
+        });
+
+        if (!voyageTeamMember)
+            throw new InternalServerErrorException(
+                `features.service.ts: voyageTeamMember not found: teamMemberId=${currFeature.teamMemberId}`,
+            );
+
+        const teamId = voyageTeamMember.voyageTeamId;
 
         await this.globalService.validateLoggedInAndTeamMember(
             teamId,
@@ -245,21 +260,20 @@ export class FeaturesService {
                         addedBy: { voyageTeamId: teamId },
                     },
                 });
-            const existingFeaturesToUpdate =
-                await existingCategoryFeatures.filter(
-                    (feature) =>
-                        feature.id !== featureId &&
-                        feature.order > currFeature.order,
-                );
-            const newFeaturesToUpdate = await newCategoryFeatures.filter(
-                (feature) => feature.order >= order,
+            const existingFeaturesToUpdate = existingCategoryFeatures.filter(
+                (feature) =>
+                    feature.id !== featureId &&
+                    feature.order! > currFeature.order!,
+            );
+            const newFeaturesToUpdate = newCategoryFeatures.filter(
+                (feature) => feature.order! >= order,
             );
             await Promise.all(
                 existingFeaturesToUpdate.map(async (feature) => {
                     await this.prisma.projectFeature.update({
                         where: { id: feature.id },
                         data: {
-                            order: feature.order - 1,
+                            order: feature.order! - 1,
                         },
                     });
                 }),
@@ -269,7 +283,7 @@ export class FeaturesService {
                     await this.prisma.projectFeature.update({
                         where: { id: feature.id },
                         data: {
-                            order: feature.order + 1,
+                            order: feature.order! + 1,
                         },
                     });
                 }),
@@ -282,13 +296,13 @@ export class FeaturesService {
                 },
             });
         } else {
-            if (order < currFeature.order) {
+            if (order < currFeature.order!) {
                 const existingFeaturesToUpdate =
-                    await existingCategoryFeatures.filter(
+                    existingCategoryFeatures.filter(
                         (feature) =>
                             feature.id !== featureId &&
-                            feature.order >= order &&
-                            feature.order < currFeature.order,
+                            feature.order! >= order &&
+                            feature.order! < currFeature.order!,
                     );
 
                 await Promise.all(
@@ -296,25 +310,25 @@ export class FeaturesService {
                         await this.prisma.projectFeature.update({
                             where: { id: feature.id },
                             data: {
-                                order: feature.order + 1,
+                                order: feature.order! + 1,
                             },
                         });
                     }),
                 );
             } else {
                 const existingFeaturesToUpdate =
-                    await existingCategoryFeatures.filter(
+                    existingCategoryFeatures.filter(
                         (feature) =>
                             feature.id !== featureId &&
-                            feature.order <= order &&
-                            feature.order > currFeature.order,
+                            feature.order! <= order &&
+                            feature.order! > currFeature.order!,
                     );
                 await Promise.all(
                     existingFeaturesToUpdate.map(async (feature) => {
                         await this.prisma.projectFeature.update({
                             where: { id: feature.id },
                             data: {
-                                order: feature.order - 1,
+                                order: feature.order! - 1,
                             },
                         });
                     }),
@@ -334,11 +348,12 @@ export class FeaturesService {
     async deleteFeature(featureId: number) {
         const currFeature = await this.findFeature(featureId);
 
-        const { voyageTeamId: teamId } =
-            await this.prisma.voyageTeamMember.findFirst({
-                where: { id: currFeature.teamMemberId },
-                select: { voyageTeamId: true },
-            });
+        const voyageTeamMember = await this.prisma.voyageTeamMember.findFirst({
+            where: { id: currFeature.teamMemberId! },
+            select: { voyageTeamId: true },
+        });
+
+        const teamId = voyageTeamMember?.voyageTeamId;
         const existingCategoryFeatures =
             await this.prisma.projectFeature.findMany({
                 where: {
@@ -346,16 +361,16 @@ export class FeaturesService {
                     addedBy: { voyageTeamId: teamId },
                 },
             });
-        const existingFeaturesToUpdate = await existingCategoryFeatures.filter(
+        const existingFeaturesToUpdate = existingCategoryFeatures.filter(
             (feature) =>
-                feature.id !== featureId && currFeature.order < feature.order,
+                feature.id !== featureId && currFeature.order! < feature.order!,
         );
         await Promise.all(
             existingFeaturesToUpdate.map(async (feature) => {
                 await this.prisma.projectFeature.update({
                     where: { id: feature.id },
                     data: {
-                        order: feature.order - 1,
+                        order: feature.order! - 1,
                     },
                 });
             }),