+++ test/utils.ts
@@ -1,5 +1,5 @@
 import { PrismaClient } from "@prisma/client";
-import { INestApplication } from "@nestjs/common";
+import { INestApplication, InternalServerErrorException } from "@nestjs/common";
 import * as request from "supertest";
 
 const prisma = new PrismaClient();
@@ -41,24 +41,51 @@ export const loginAndGetTokens = async (
 };
 
 export const getNonAdminUser = async () => {
-    try {
-        const adminRole = await prisma.role.findUnique({
-            where: {
-                name: "admin",
-            },
-        });
-        return prisma.user.findFirst({
-            where: {
-                roles: {
-                    none: {
-                        roleId: adminRole.id,
-                    },
+    const adminRole = await prisma.role.findUnique({
+        where: {
+            name: "admin",
+        },
+    });
+
+    if (!adminRole)
+        throw new InternalServerErrorException(
+            "test/utils.ts: admin role not found in the database",
+        );
+
+    const adminUser = prisma.user.findFirst({
+        where: {
+            roles: {
+                none: {
+                    roleId: adminRole.id,
                 },
             },
+        },
+    });
+    if (!adminUser)
+        throw new InternalServerErrorException(
+            "test/utils.ts: admin user not found",
+        );
+    return adminUser;
+};
+export const getUseridFromEmail = async (
+    email: string,
+): Promise<string | undefined> => {
+    try {
+        const user = await prisma.user.findUnique({
+            where: {
+                email,
+            },
+            select: {
+                id: true,
+            },
         });
+        if (!user) {
+            throw new InternalServerErrorException(
+                "test/utils.ts: user not found",
+            );
+        }
+        return user.id;
     } catch (e) {
         console.log(e);
-    } finally {
-        await prisma.$disconnect();
     }
 };